feat: build dav1d

+++ Scripts/ci.sh
@@ -1,19 +1,41 @@
 #!/bin/sh
 
+
+# build old
 rm -rf tool
 git clone -b tool --depth 1 https://github.com/zyuanming/FFmpeg-iOS tool
 cd tool
-
 swift run
+cd ..
+
+
+rm -rf ffmpeg-kit-main
+git clone -b main --depth 1 git@github.com:arthenica/ffmpeg-kit.git ffmpeg-kit-main
+cd ffmpeg-kit-main
+sh ios.sh --enable-dav1d --disable-armv7 --xcframework
+
+
+dav1d_arm64=prebuilt/apple-ios-arm64/dav1d/lib/libdav1d.a
+dav1d_arm64_headers=prebuilt/apple-ios-arm64/dav1d/include
+xcodebuild -create-xcframework -library $dav1d_arm64 -headers $dav1d_arm64_headers -output prebuilt/bundle-apple-xcframework-ios/libdav1d.xcframework
+
+cd ../tool
+
 
 # TAG=v0.0.6-`date +b%Y%m%d-%H%M%S`
 TAG=0.0.6
 
 cp ../Package.swift .
 
-for f in Frameworks/*.xcframework
+for f in prebuilt/bundle-apple-xcframework-ios/*.xcframework
 do
 	f=`basename $f .xcframework`
+
+  # 检查文件名是否为ffmpegkit.framework
+  if [ "$f" == "ffmpegkit" ]; then
+      # 如果是ffmpegkit.framework，则跳过
+      continue
+  fi
 	echo $f...
 	rm Package.swift.in
 	mv Package.swift Package.swift.in
@@ -57,14 +79,14 @@ gh release create -d $TAG -t "FFmpeg-iOS $TAG" --generate-notes --verify-tag
 
 echo "Uploading Binaries..."
 
-XCFRAMEWORK_DIR=tool/Frameworks
-for f in $(ls "$XCFRAMEWORK_DIR")
-do
-    if [[ $f == *.zip ]]; then
-        gh release upload $TAG "$XCFRAMEWORK_DIR/$f"
-    fi
-done
+# XCFRAMEWORK_DIR=tool/Frameworks
+# for f in $(ls "$XCFRAMEWORK_DIR")
+# do
+#     if [[ $f == *.zip ]]; then
+#         gh release upload $TAG "$XCFRAMEWORK_DIR/$f"
+#     fi
+# done
 
-gh release edit $TAG --draft=false
+# gh release edit $TAG --draft=false
 
 echo "All done!"
